Vmectrace now working
